services/problem-enricher/requirements.txt
anthropic>=0.18.0
openpyxl>=3.1.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
rich>=13.0.0
